LayoutTests/fast/forms/month-multiple-fields/month-multiple-fields-preserve-value-after-history-back.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<div id="container">
<input type="month" id="test1">
<input type="month" id="test2" min="1999-01" max="1999-12">
</div>
<iframe id="iframe"></iframe>
<script>
description('Checks input value is preserved after backward and forward');
if (!window.eventSender)
    debug('Please run within DRT/WTR');
window.jsTestIsAsync = true;

var iframe = document.getElementById('iframe');
var testInput1 = document.getElementById('test1');
var testInput2 = document.getElementById('test2');
var state = 'direct';
var pageName = location.search || '?page0';
var testCases = [];

function back(newState) {
    parent.state = newState;
    history.back();
}

function failed(message) {
    testFailed(message + ' in ' + state);
    finishJSTest();
}

function fieldsText(input) {
    if (!window.internals)
        return "";
    var editElement = window.internals.oldestShadowRoot(input);
    return editElement.textContent;
}

function finishTest() {
    runTestCases();
    document.body.removeChild(document.getElementById('container'));
    finishJSTest();
}

function forward(newState) {
    parent.state = newState;
    history.forward();
}

function keyDown(key, modifiers) {
    if (!window.eventSender)
        return;
    eventSender.keyDown(key, modifiers);
}

function runTestCases() {
    for (var index = 0; index < testCases.length; ++index) {
        var testCase = testCases[index];
        debug(testCase['description'] + ':');
        shouldBeEqualToString('"' + testCase['actual'] + '"', testCase['expected']);
    }
    debug('');
}

window.onload = function() {
    switch (pageName) {
    case '?page0':
        iframe.src = location + '?page1';
        break;

    case '?page1':
        switch (parent.state) {
        case 'direct': {
            testInput1.value = '12012-10';
            testInput2.focus();
            keyDown('upArrow'); // -> [October] 12012
            var nextUri = location.toString().replace('?page1', '?page2');
            setTimeout(function () { window.location.href = nextUri; }, 0);
            break;
        }
        case 'backTo1':
            parent.testCases.push(
                {
                    'description' : 'Went back to a page. Checking a form control which had a full value',
                    'actual' : testInput1.value,
                    'expected' : '12012-10' },
                {
                    'description' : 'Went back to a page. Checking a form control which had a partial value',
                    'actual' : fieldsText(testInput2),
                    'expected' : 'January ----'
                });
            forward('forwardTo2');
            break;
        case 'backTo1Second':
            parent.finishTest();
            break;
        default:
            failed('Unexpected state(' + parent.state + ')');
            break;
        }
        break;

    case '?page2':
        switch (parent.state) {
        case 'direct':
            testInput1.value = '0001-11';
            testInput2.focus();
            keyDown('\t'); //        -> --------- [yyyy]
            keyDown('downArrow'); // -> --------- [1999]
            back('backTo1');
            break;
        case 'forwardTo2':
            parent.testCases.push(
                {
                    'description' : 'Went forward to a page. Checking a form control which had a full value',
                    'actual' : testInput1.value,
                    'expected' : '0001-11'
                },
                {
                    'description' : 'Went forward to a page. Checking a form control which had a partial value',
                    'actual' : fieldsText(testInput2),
                    'expected' : '--------- 1999'
                });
            back('backTo1Second');
            break;
        default:
            failed('Unexpected state(' + parent.state + ')');
            break;
        }
        break;

    default:
        failed('Unexpected page(' + pageName + ')');
        break;
    }
};
</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
