LayoutTests/platform/chromium-win/fast/forms/month-multiple-fields/month-multiple-fields-preserve-value-after-history-back-expected.txt

Checks input value is preserved after backward and forward

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Went back to a page. Checking a form control which had a full value:
PASS "12012-10" is "12012-10"
Went back to a page. Checking a form control which had a partial value:
FAIL "January, ----" should be January ----. Was January, ----.
Went forward to a page. Checking a form control which had a full value:
PASS "0001-11" is "0001-11"
Went forward to a page. Checking a form control which had a partial value:
FAIL "---------, 1999" should be --------- 1999. Was ---------, 1999.

PASS successfullyParsed is true

TEST COMPLETE
